For classes, add a field to the script object just like functions (#149)
Then, (external) classes can be found by the PA when they are imported. Otherwise, they cannot be found.

We can't a test for this because we found that `com.ibm.wala.cast.python.test` is run twice during the Maven build. And, the behavior of the code differs between the builds. Thus, on one run, the assertions pass, and on the other, the assertions fail. The only difference I can see is that the first test uses Python files from `target/`, while the second passes uses the ones from the repo. Other than that, I have no idea what the difference is and why it's run twice in the first place.

Still, this fix works in our client application. Thus, I feel that the fix would be beneficial for someone as it solves a significant bug in our client.