Bug fix and other updates (#173)
These are some smaller updates; a larger one is coming soon but wanted to keep up-to-date with the smaller ones.

Most significant change is: [Fix bug with determining whether a module is "local."](https://github.com/wala/ML/commit/0f6872dda7bdcffdfedd884f794a366481dff0a0)

The old code didn't consider the full path of the module, which is what is contained in `localModules`. But, it still doesn't do what I think it
should do, which makes me believe that the generated CAst is incorrect. A fix is forthcoming.