Add `PYTHONPATH` to the analysis (#91)
* First test.

* Add comment.

* Test 2.

Should fail.

* More tests.

* Add comment.

* Extract local variable refactoring.

* Improve variable name.

* Add logging.

* Decrease logging level.

* Apply spotless.

* Fix bug with determining whether a module is "local."

The old code didn't consider the full path of the module, which is what
is contained in `localModules`. But, it still doesn't do what I think it
should do, which makes me believe that the generated CAst is incorrect.

* Replace slash with dot in test filenames.

* New test.

This one has a sibling script in the module that uses the other sibling.

* Implement `getSignature()` for `com.ibm.wala.cast.python.parser.PythonParser.translateToCAst().PythonScriptEntity<T>`.

No signature currently so return `null`.

* Actually add the test files.

* Fix filename for test.

* Progress.

* Progress.

* Cleanup.

* Remove redundant log.

* Progress.

* Fix compilation error.

* Progress.

* Progress.

* Progress.

* Progress.

* Only test PYTHONPATH for Jython3.

It's only supported for Jython3.

* Add docs.