Continue module parsing despite Jython parsing errors (#206)
Let's see if we can still go through (similar to
[b3c28a86a282f22f45b8e1203ef8eec8dc91214a](https://github.com/wala/ML/pull/125)).

## Differences
1. Log a warning instead of throwing an exception, which stops the
module from processing. We are already doing this further up the stack
anyway:
https://github.com/wala/ML/blob/ef1cca799d7726bfc70cec7b838ebeec2516d5e4/com.ibm.wala.cast.python/source/com/ibm/wala/cast/python/client/PythonAnalysisEngine.java#L159-L166
2. Continue to convert to CAst despite the parsing error coming out of
Jython. This is a workaround for https://github.com/wala/ML/issues/205.
I think this should be OK (best-effort) as I don't see it contributing
to any soundness issues. And, from what I am seeing in the
[logs](https://github.com/ponder-lab/ML/actions/runs/9897547494/job/27342295209#step:12:32780),
the parsing errors are "mild."