Deal with multiple possible callables (#121)
* Return `null` when there are multiple possible callables.

* Add test to exercise call string imprecision.

Based on the call string length. See https://github.com/wala/WALA/discussions/1417#discussioncomment-10085680.

* Expect the test to fail.

In the past, we could add 0's to the parameters, but since we are not
enforcing the existing of the node in the CG, we can no longer do that.
Still, this test should now fail if https://github.com/wala/ML/issues/207 is fixed.